Replace UnifiedCacheManager.ts and ICacheStorage.ts with CacheManager.ts
