+++ lib/msal-browser/src/app/PublicClientApplication.ts
@@ -530,7 +530,7 @@ export class PublicClientApplication {
      * @returns {@link IAccount[]} - Array of account objects in cache
      */
     public getAllAccounts(): IAccount[] {
-        return SPAClient.getAllAccounts(this.browserStorage);
+        return this.browserStorage.getAllAccounts();
     }
 
     /**
