+++ lib/msal-common/src/index.ts
@@ -18,9 +18,7 @@ export { B2cAuthority } from "./authority/B2cAuthority";
 export { AuthorityFactory } from "./authority/AuthorityFactory";
 export { AuthorityType } from "./authority/AuthorityType";
 // Cache
-export { ICacheStorage } from "./cache/interface/ICacheStorage";
 export { CacheManager } from "./cache/interface/CacheManager";
-export { UnifiedCacheManager } from "./cache/UnifiedCacheManager";
 export { AccountCache, AccessTokenCache, IdTokenCache, RefreshTokenCache, AppMetadataCache } from "./cache/utils/CacheTypes";
 export { CacheHelper } from "./cache/utils/CacheHelper";
 export { Credential } from "./cache/entities/Credential";
